Update 5-exposure-radiogenic-risk-activities.md
Added concurred edits. Danny -- there's a nested bulleted list that will need your attention. Also -- a link to a document that doesn't, best as I can tell, live on a VA site in a way we can link to it. We do need to link to the content, though. Can you stub out a page for this content in the Info Page Template so we can link? Thanks!