Separate out VOA HTML into HTML fragments.
Make the structure of the page more apparent and the content far more
manageable.

Also puts the primary navigation into a set of vertical tabs and removes
the secondary navigation so it looks like a bunch of cards on one page.

Checking site.domain to insert a stub page when rendering in production.