Use a CSP to disable external resource loads.
PhantomJS's slowdown is caused by external resource loads to things like
Google analytics and usasearch which are embedded in the headers and
footers of this site. It causes a runtime of > 4s per test due to lack of
cacheing. It also breaks hermeticity of the test. These modules should
not be required for most of our tests.

Using a CSP header that enforces a same-origin only policy for *all* resource
loads provides a way to speed up tests w/o source code changes. It means
tests for things depend on external resources will break, but maybe
there needs to be an alternate testing strategy (mocks, etc) for those anyway.