Add build_production rake task.
Currently, deploys are done off the exact same build settings in staging
and production which is allowing for code leakage to the production
website. This adds a `build_production` rake task that produces a
_site_production directory.  Having a different output target makes
it harder to cross wires during deploy.

This isn't a complete solution as the CI set up does NOT run against
the production directory. Checking this is first to fix the current
production fire. Will fix CI next.