Move healthcare-app JS into webpack.
This allows for es6 transpilation, and more proper control of
deployment w/o hacking on jekyll load the script as a big
inline blob into the html.

Code changes to _health-care/_js/_form.js are mostly
  * Using es6 module exports instead of a hacked "module" pattern
    with a function.
  * Replacing "var" with "let" because, declarazing hoisting is a
    batsh** crazy and stupid language feature.
  * Requiring the needed libraries.
  * Moving the foundation intialization into the app entrypoint.

In addition, a basic client side router is added into entry.js so
the health-care app javascript bundle is not loaded unless it's on the
right page.

Unforunately, because the client side router us based on looking at
window.location, the karma iframe serving infrastructure also needed to
be modified so the test fixture would invoke the javascript. There's
probably a cleaner way to fix this, but the current hack seems most
expedient.