Use babel-runtime. Use special env for hot reload.
Using the Babel Runtime Transform shares the adapter functions created
for ES6 polyfills reducing the final code size.

Hot load should only be used in interactive development. Switching
the environment to a special name ("hot") makes it easier to only
enable it when it's wanted like in the `npm run watch` target.