Make ErrorableSelect options accept array.
Previously the ErrorableSelect option prop took a JS object which
was used to map the (key, value) pairs into (value, label) on an
<option> element.

This has the undesirable effect of requiring preventing values from
having spaces or being JS keywords.

Change the API so that options is an array of either strings, or objects
with an explicit {label:, value:} structure.