Navigate by Sections. Remove Panel concept.
Wire up the side-nav to navigate to each section. With this Panels are
an obsolete concept and all Panels have been deleted.

The model data is now located by parsing the current URL to find the
relevant state for a given section. This is handled in <HealthCareApp>.

Routes are extracted into a separeate file for easier use in unittests.

Still missing some tests, but pushing now to unblock others.