Add in broken link checking. Fix pension pages.
Broken link checking is super important for CI.

The pension pages on the otherhand look like they have been broken for a
while. The fixes here at least make the text visible and allow the
broken link checker to pass but #2732 has been filed to figure out what
to do with these pages.