E2E and accessibility tests use prod build type properly
The BUILDTYPE env is used to determine which content the test
server serves during e2e and accessibility tests. While the production
build type was used for the build command, the BUILDTYPE defaulted
to development, causing the test server to serve content that didn't
exist, which in turn caused these tests to fail.

This change sets BUILDTYPE appropriately.