Addressing layout and organization issues
Further reorganization to get the most important info at top and to use plain language. It is tricky to find another way to say "clinician" other than "doctor or other medical professional," which is a mouthful. I opted for using "clinician" once and "doctor" thereafter. The issue being that it won't necessarily be a doctor--could be a PA or some other assistant.