PL Edit: Disability_Cond_Exp_Contaminated Water
Plain Language Edits for vets.gov/disability-benefits/conditions/exposure-to-hazardous-materials/contaminated-drinking-water-at-camp-lejeune/

Question for Beth: Should there be a link to the online application here? I don't see it on live site, but since we talk about it, it seems that we'd want to have a link to it. Thoughts?

Note: Same issue with some content not appearing in preview.