[WIP] PL Edit: Disability_Cond_SpecClaims_Pre-Stab
Plain language update for: /disability-benefits/conditions/special-claims/prestabilization/

Ready for Beth's review.
Note for Beth: This page was missing some of the standard info sections that we've been including. The corresponding page on va.gov seemed to have this info so I added here for consistency. Please let me know if we don't want to include.