[WIP] Disability_Apply_OneYear
Plain language update for: /disability-benefits/apply-for-benefits/one-year/

Ready for Beth's review. Question for Beth: This one had "Disability" as a separate benefit. I've removed it because I believe all of these benefits fall under "disability benefits" and that hasn't been called out separately in any of the others. Please let me know if that's incorrect to remove. Thanks!