[WIP] Disability_ClaimsProc_ClaimTypes_NewClaim
Plain language update for: /disability-benefits/claims-process/claim-types/new-claim/

Ready for Beth's review. Question for Beth: I believe my changes are still accurate, as it appears that "new claim" references a new claim for an existing covered disability (and not a new claim in general) since the bullets speak to "more" and "a shift," but please double check me on this.