[WIP] PL Edit: Disability_ClPr_ClTy_Pre_QuickStart
Plain language update for: /disability-benefits/claims-process/claim-types/predischarge-claim/quick-start/

Ready for Beth's review. Beth, a question for you: this is first one I think where I've seen the "available benefits" like this. should we ask if this is what it should read or if it should speak to financial support like other pages?