Travis performance enhancements for workflow
Introduces two performance enhancements for our Travis configuration
based on our workflow. Travis builds will not be run for pushes to
master or production, and the content team may develop on branches
specifically designed to operate preview builds and not run tests.

For the general workflow:

1. Pull requests will trigger Travis to run all non-accessibility
   tests.
2. Pull requests to production will run all tests, including the
   accessibility tests.

For the content team workflow:

1. Branches with the name `content/wip/X`, where X can be a brief
   description of the feature, will fail all Travis builds.
2. These branches will only trigger a build of the development
   BUILDTYPE, since this is the build used for the Heroku review app.
3. These branches will trigger Heroku review app builds, allowing
   content preview while they're in public.
4. A new branch must be created from `content/wip/X` for a new Pull
   Request, where the tests will be run. This can be done when content
   development is completed.

Travis runs two different types of builds - PR and push. Push builds
test the branch itself, and PR builds test the resulting build into the
target. These may be different, since the push may not yet contain updates
to the target branch. Our configuration requires running three builds
(development, staging, and production), which results in a total of
6 builds (each environment x (PR + push)):

A build success is required to merge code from a PR. On merge, Travis
will run a push build for the result on the target branch. For
a typical feature branch merge into production before this update,
this results in a total of 9 builds - 6 for the PR, and 3 for the push
after the merge.

For merges to master:

| Type | Branch      | BUILDTYPE   | E2E | Accessibility | Deploy      |
| ---  | ---         | ---         | --- | ---           | ---         |
| push | `feature/X` | development | yes | no            | no          |
| push | `feature/X` | staging     | yes | no            | no          |
| push | `feature/X` | production  | yes | no            | no          |
| PR   | `feature/X` | development | yes | no            | no          |
| PR   | `feature/X` | staging     | yes | no            | no          |
| PR   | `feature/X` | production  | yes | no            | no          |
| push | `master`    | development | yes | no            | development |
| push | `master`    | staging     | yes | no            | staging     |
| push | `master`    | production  | yes | no            | no          |

Total: 9 x E2E

For merges to production, which must run the accessibility suite:

| Type | Branch       | BUILDTYPE   | E2E | Accessibility | Deploy     |
| ---  | ---          | ---         | --- | ---           | ---        |
| push | `feature/X`  | development | yes | no            | no         |
| push | `feature/X`  | staging     | yes | no            | no         |
| push | `feature/X`  | production  | yes | no            | no         |
| PR   | `feature/X`  | development | yes | no            | no         |
| PR   | `feature/X`  | staging     | yes | no            | no         |
| PR   | `feature/X`  | production  | yes | no            | no         |
| push | `production` | development | yes | no            | no         |
| push | `production` | staging     | yes | no            | no         |
| push | `production` | production  | yes | yes           | production |

Total: 9 x E2E + 1 x Accessibility

Since PRs require an updated master prior to merge, we can safely
assume that multiple PRs will not be merged into master at the same
time and result in a master branch state that was untested. So, we can assume
that the PR tests the resulting target branch push and that a new test
is unnecessary.

For merges to master we can:

| Type | Branch      | BUILDTYPE   | E2E | Accessibility | Deploy      |
| ---  | ---         | ---         | --- | ---           | ---         |
| push | `feature/X` | development | no  | no            | no          |
| push | `feature/X` | staging     | no  | no            | no          |
| push | `feature/X` | production  | no  | no            | no          |
| PR   | `feature/X` | development | yes | no            | no          |
| PR   | `feature/X` | staging     | yes | no            | no          |
| PR   | `feature/X` | production  | yes | no            | no          |
| push | `master`    | development | no  | no            | development |
| push | `master`    | staging     | no  | no            | staging     |
| push | `master`    | production  | no  | no            | no          |

Total: 3 x E2E

And for merges to production:

| Type | Branch       | BUILDTYPE   | E2E | Accessibility | Deploy     |
| ---  | ---          | ---         | --- | ---           | ---        |
| push | `feature/X`  | development | no  | no            | no         |
| push | `feature/X`  | staging     | no  | no            | no         |
| push | `feature/X`  | production  | no  | no            | no         |
| PR   | `feature/X`  | development | yes | no            | no         |
| PR   | `feature/X`  | staging     | yes | no            | no         |
| PR   | `feature/X`  | production  | yes | yes           | no         |
| push | `production` | development | no  | no            | no         |
| push | `production` | staging     | no  | no            | no         |
| push | `production` | production  | no  | no            | production |

Total: 3 x E2E + 1 x Accessibility

This results in a nearly 3 times improvement of the total CI build time
required to get code into master, and nearly that for build time to get
code into production.