Update flight-training.md
"If flight training is part of your college degree at an approved college or university, we pay the required tuition and fees, a monthly housing allowance, and a books and supplies stipend." Is this sentence only related to the post 9/11 bill or also REAP and MGIB? If both, leave as above. If just post 9/11, pull into that bullet.

"If your school is a public college or university, VA reimburses the school the full cost of tuition and fees*. There is no yearly limit." --> confirm the last sentence is true (I added it for consistency--if all items on the list have a yearly limit, you can pull it to the top so it's not repeated in each bullet

In the first section it's "we pay" and in the rest it's "VA pays"... consider making consistent

The section "### Can I get benefits for vocational flight training and still have GI Bill benefits left?" doesn't make sense to me--will it make sense to vets who understand the program? I'm not sure the question used for the header is quite right.