Enable tests for pushes to master and production
For merge after "Require branches to be up to date before merging" is unchecked.

For performance purposes we'd disabled the requirement to run tests on pushes to
master and production branches, since the PR containing the merge was fully tested
prior to merge. With this option disabled, we'll re-enabling these tests.