Auto-hide #va-modal
Adds CSS rule to hide #va-modal to be opened if '#modal' is found in the url (content/includes/modal.html).

I wasn't sure if calling modal.hide() would be a better solution if #va-modal is used elsewhere, so I kept that as a comment in modal.html.

I looked for other uses of both #va-modal and modal.html, but couldn't find any...but at this point, I still don't know if it's safe, so somebody more familiar with the codebase should probably do some pretty thorough testing.

Also, Atom likes to remove trailing spaces, so that's what the seemingly-unchanged lines are.