Final PL proof 2: 4-action-page-questions
We had two links to the fact sheet (a pdf) from the C-123 page: one in the top right box called Need help? that was in this file, and one under Get more information at the bottom of the C-123 page. I removed the former since it's not really giving help, it's giving info.