First pass PL
Beth: I think for most of these, it's okay to have the link within the sentence because it starts the sentence off and is a program name. The exception is the last one, where "a monthly subsistence allowance" was a link right in the middle of the sentence. So I revised that to be "See the current subsistence allowance rates" as a 2nd sentence. Also changed "You can earn" to "You may also be able to earn" because the VA.gov page this goes to says, "In some cases"...so I think "can" would be over-promising there.