Updates - Ready for Beth
Beth: I moved the rates question up because, with the newly proposed approach to the rates table, there would be accordions within the question...and therefore, we wouldn't want an accordion for the question as well. So this would be best up with the non-accordion questions. We'll need to wait for Mel's input on the approach to the table. Thanks!