Adding descriptions to resources major links
Peggy & Beth: I do think it would be best to add descriptions here so it's very clear what each resource provides. I also updated the title of the "Job Search" resource to align with the page title on that page. Note that the links to these page do not currently work. So we'll need to figure that out before launch!