Updating link based on disability updates
redirecting /disability-benefits/conditions/service-connected/ to /disability-benefits/conditions/. This change is per our request for the removal of the disability-benefits/conditions/service-connected/ page (it was a repetitive page that wasn't in the actual nav before, but got swept into it in the new template changes...I've asked Mel to remove it). Thanks!