conditional content updates
Note for Peggy: I used "at least one of these descriptions is true for that Veteran" instead of "that Veteran meets at least one of the requirements listed below" for the one question above because the description language is a little softer and more empathetic for a case in which death, being a POW, etc. is the requirement. Just wanted to make sure to let you know that as you're editing.