526 update upload mg (#8095)
Update the schema for uploads to give vets-api the extra information it needs about each uploaded file. Before, we weren't sending anything up with the submit as vets-api was storing uploads against a user. Now, we're sending up an array of GUIDs (as confirmationCode), filename, and document type for each uploaded record.