Change handling of school addresses from the search tool
- Combine the second and third address fields so no data is lost; this
is especially important for some international addresses that the search
tool returns as the many of those contain important address info in the
`address3` field.
- If the address from the search tool is domestic, transform the country
to the prop enum value, "United States".
- If the address is not domestic, send the country as-is and add the
`viaSearchTool=true` field on the `address` so the backend will use the
school address schema that accepts any string for the country value.