pull up properties a level and transform back down a level (#9568)
https://app.zenhub.com/workspaces/vft-59c95ae5fda7577a9b3184f8/issues/department-of-veterans-affairs/vets.gov-team/16747

On bah-0994-schema-changes, the City and State fields are appearing as expected and have "required" next to them, but if the fields are left blank, the user can enter a start date, click continue and go on to the next page. If the "in-person" or "online and in-person" radio is selected and no information is entered, the "please provide a response" message shows for the start date field, but not City or State.