Polyfill fetch for all unit tests
Before making this fix, unit tests that expected Response to be defined
woudl fail if you ran them in isolation. But they did pass just fine if
you ran all unit tests. Turns out the problem was that the vic-v2 unit
tests imported the `isomorphic-fetch` package and that was the reason
that Response-dependent tests would pass if you ran them all. This
change adds `isomorphic-fetch` to the list of packages required by
Mocha. Thus any unit test that expects fetch-related classes to be
defined (in our case that's just the Response class) will now pass when
run in isolation.