Add support for two new CTA widget types (#10593)
With this change users of the VA.gov CMS can choose which type of
appointment component/widget to embed. Previously the only way to embed
an appointment widget for scheduling/cancelling appointments was to
embed two appoint widgets on the page and the FE code would make the
first widget a "view" appointments widget and it would make the second a
"schedule" appointments widget.