Remove support for `appointments` widget type
Also remove support configuring widgets differently based on their
position/index on the static page. This feature was only used in this
one case; to treat appointments widgets differently depending on their
position on a page. This is no longer required because CMS content
editors can use either the `view-appointments` or
`schedule-appointments` widget type IDs to directly control the
configuration of the CallToActionWidget that gets embedded in the static
page