AC 4
When the checkbox is checked:
the ZIP code question (AC # 2) and input box is hidden.
the benefit estimator is updated to display the international housing VA national average rate when the answer to "Did you use your Post-9/11 GI Bill benefits for tuition, housing, or books for a term that started before January 1, 2018?" is yes.
the benefit estimator is updated to display the international housing DOD national average rate when the answer to "Did you use your Post-9/11 GI Bill benefits for tuition, housing, or books for a term that started before January 1, 2018?" is no.