AC 7
In the "School Locations" section, international locations display the international housing rate in the "Estimated Housing" column.
the international housing VA national average rate when the answer to "Did you use your Post-9/11 GI Bill benefits for tuition, housing, or books for a term that started before January 1, 2018?" is yes.
the international housing DOD national average rate when the answer to "Did you use your Post-9/11 GI Bill benefits for tuition, housing, or books for a term that started before January 1, 2018?" is no.