4422 global downtime notification (#11427)
* add downtime behavior

* add downtime behavior

* use feature toggle

* revert change to burials form config

* remove exists check from reducer

* add forking to backend statuses reducer

* Fix failing action unit test

The updated getScheduledDowntime return function now expects a state
argument to be there, and the apiRequest helper also expects the
resposne to contain a `url` otherwise it will break.

* Fix getBackendStatuses unit tests

The function now checks a state variable and also makes a call to
dispatch to indicate if global downtime is active

* Update tests for external services reducer

If the action specifies that the global downtime is active then the
status will change

* Set state in action test to be empty objects

It doesn't matter what the state is as long as it exists.  Maybe it will
be worth using something like optional chaining instead of forcing
something to exist.

* Also set state to be empty here like last commit

* Add tests for createGlobalMaintenanceWindow helper

* Refactor & add unit tests for getBackendStatuses

getBackendStatuses has been refactored to take a downtimeWindow as an
argument with the default being the hardcoded scheduled values.  Having
the argument allows us to pass in differen values for testing.

* Properly mock apiRequest

* Fix moment issues within tests

* Use .catch() in promise chain instead of res.ok

The apiRequest helper doesn't return the response status but it does
reject a promise if the response isn't ok.

I also think we may be able to remove the try/catch wrapper but I'm
leaving it in for now just because.

* Try to mock fetch a different way

This is a mess.  I'm trying to add another test to verify that
ERROR_SCHEDULE_DOWNTIME is dispatched when a fetch response isn't ok, so
to do that I will need to be able to mock fetch so that it can succeed
in one test and "fail" in another.

* Clean up a bit & reset the dispatcher

* Add test to cover fetch response not being ok

* Invoke downtime when response has error data

We are doing this because sometimes the api will respond with a 200 even
when there is a server error.

* Adjust global downtime banner on sign-in modal

Disabling the buttons will come next

* Disable the buttons on global downtime

This duplicates some code in ExternalServicesError - it may be worth
cleaning it up a bit

* Fix some tests

Destructuring caused the function to be called apparently, and we are
only looking for it to be called once.

Additionally, sometimes the statuses prop doesn't exist so we need a
null check for that using optional chaining

* add downtime notification to gibct; update default downtime message for apps

* Add tests for the ExternalServiceError callback

The component now has a callback prop for when it renders.  This allows
parent components to know when it is rendered, which means that the
parent components can know when a service is down.  This is used to
disable login buttons

* Add global dependency to 0993-edu-benefits

* Add to 0994

* Add test for downtimeWindow data

I feel like this could be done a bit better, but it verifies that when
downtime happens the global downtime data is the right shape

* Add to 996

* Add global dependency to 1990

* Add global dependency to 1990e

* Add global dependency to 1990n

* Add global dependency to 1995

* Skip weird test for now

This fails on Jenkins and when `npm run test:unit` is run locally, but
if the specific file is run then it passes. Due to the time crunch I
don't think this is worth investigating at the moment

* Add global dependency to 5495

* Add downtime to 686's intro page

* Added global dependency to chapter31-vre

* Add global dependency to chapter36-vre

* Add global dependency to feedback-tool

* Add global dependency to veteran-representative

* Add global dependency to vic-v2

* revert downmessaging removal, remove feature flag

* add localhost check; fix linting issues

* disable unit test

* disable e2e test for downtime notification

* add correct downtime window

* fix backend statuses api failure action

* update time for test

* update time

Co-authored-by: Brooks Johnson <2008881+bkjohnson@users.noreply.github.com>
Co-authored-by: Chris Valarida <chris.valarida@adhocteam.us>