Add node-health_care_local_health_service
Note: The test data still needs to be modified, but I'm going to work on
the child entity's transformer before looping back around to this one.
Hopefully, this will catch any discrepancies between what we'll actually
be getting and what the templates will actually be using.

This is the price we pay for not being able to just use the test data
as-is—we need to modify it a littl bit, but there's no good way to
compare it to the original test data.