Filter transform press release (#11505)
* Run generator

* Fix some type issues in raw schema

* Make some changes to the transformer

* Remove wrapping "entity"

This looks like a lot but it's really just the outer entity key being
removed and the indentation being updated

* Fix the press release contact field

* Add metatags

* Make some schema & transformer updates

Vague, I know.  This adds the `processed` prop for one of the keys to
the schema and transformer in addition to setting the entityPublished
value

* Adjust the test data

Usual `\r` stuff, but the metatags in our tome-sync export have "VA.gov
CMS" as part of the title rather than "Veterans Affairs"

* Add fieldPressReleaseBlurb to the region page

This modifies a child entity's schema and transformer in addition to
adding the extra data from that child entity to the test data for the
parent.  This is to keep things simple and consistent.

* Properly transform the release date

* Improve schemas/transformers to fix content build

* Add press release blurb to health care region page

* Use some common schemas

* Change press release downloads to entity ref array

This might not matter since I haven't seen it be anything other than an
empty array