Try running a "build" before getting PR number
The docs [1] say that CIRCLE_PR_NUMBER and CIRCLE_PULL_REQUEST will be
available as environment variables, but based on a previous [2] that
doesn't seem to be the case.  The docs say that the environment
variables are:

> exported in each build

so maybe that means that they are only exported if the specific "build"
job gets run in a workflow.

[1] - https://circleci.com/docs/2.0/env-vars/#built-in-environment-variables
[2] - https://app.circleci.com/jobs/github/department-of-veterans-affairs/vets-website/19/parallel-runs/0/steps/0-102