Don't add review comments for the same problems
Identifying the review comments that the "bot" left previously is a bit
tricky.  First, we fetch all the comments on a PR and ignore the ones
that are outdated (identified by having a null position).

This part is a bit hacky and I haven't thoroughly tested it yet - we
also ignore the comments that weren't made by a login equal to the
BOT_NAME environment variable.

This means that the environment variables for the token and the name
should be for the same account.

Now that we have only the comments that our review bot left, we can
determine if we have any new additions to make.  A new addition is an
addition which is on a new path/position combo and that has a new body.