Add initial draft for icon-check job
This is for accessibility reasons.  FontAwesome icons typically use the
<i> tag to add icons, and sometimes icons are used to convey meaning
that isn't captured in text.  This is what we want to be on the lookout
for.  In the cases where icons are used with a semantic purpose in mind,
we want to make sure that those semantics are conveyed to a screen
reader.