Flag PR for manual review on certain additions (#11670)
* See if circle sets PR number

* Tweak syntax

* Setup a workflow to run a pr-check on non-master

* Indent config

* Just try to run the job in the workflow

* Try running a "build" before getting PR number

The docs [1] say that CIRCLE_PR_NUMBER and CIRCLE_PULL_REQUEST will be
available as environment variables, but based on a previous [2] that
doesn't seem to be the case.  The docs say that the environment
variables are:

> exported in each build

so maybe that means that they are only exported if the specific "build"
job gets run in a workflow.

[1] - https://circleci.com/docs/2.0/env-vars/#built-in-environment-variables
[2] - https://app.circleci.com/jobs/github/department-of-veterans-affairs/vets-website/19/parallel-runs/0/steps/0-102

* Use list syntax for requires

* Include colon

* Try to run eslint-disable check in circle

Eventually we would want to limit this to happen only on branches with
PRs attached

* Remove build from workflow

* Add an eslint-disable

* Remove the trailing colon

* Checking for checks ;)

* Try to leave a comment on each offending line

* Install dependencies before running script

* Try setting comments by offset instead

This will need some more explanation, but this is due to how the API
works

* Set position to be 1 instead of 0

* Use a map instead of a separate function

* Delete old commented out code

This includes the pasted github actions code that was used as an example

* Create function to encapsulate some things

This will help when we want to get line additions that aren't for
disabling eslint

* Specify pattern from CI job

* Update position even for deleted lines

This will help us use the right position when making the review comment
based off of position.  If the API changes so that we can use line
instead we will have to undo this.

* Make review comments configurable

* Add install-dependencies job as requirement

We can eventually use this for other workflows

* Add install-dependencies to workflow

* Go back to installing dependencies as a step

We can optimize this later.  This ended up breaking because the job
didn't have access to the code, so apparently `checkout` doesn't make
code available between jobs

* Get the diff from the Github API

This may make more sense as the code gets tweaked.  Also, it feels more
"node-y" to use a REST API instead of using something cobbled together
using child_process

* Refactor job so to be parameterized

This should allow us to use the same job to check for different things
in PRs

* Turn jobs into mappable objects in config

* Undo changes for parameterized jobs

This only seems available for version 2.1 - Not sure if we want to use
that yet

* Add job to check for Sentry

* Make Sentry call

* Trigger Circle

* Try to fix position reset

This is from Github's API documentation on the position for PR review
comments:

> The position value equals the number of lines down from the first "@@"
> hunk header in the file you want to add a comment. The line just
> below the "@@" line is position 1, the next line is position 2, and so
> on. The position in the diff continues to increase through lines of
> whitespace and additional hunks until the beginning of a new file.

Since the position continues to increase until a new file, I'll try
only resetting it when we reach a line that starts and ends with @@

* Make sure position is updated on new chunk

If we reach the double @@ line in the same file, make sure position gets
updated

* Translate add_lines from bash to JS

* Remove grep child_process

* Fix bug with position not resetting

* Refactor for promise chaining

* Remove a Sentry call

This is to test if the script still leaves a comment on the diff

* Temporarily disable eslint check

* Only include additions in output

This could have been achieved in the filtering step by prepending the
CODE_PATTERN search with a regex to match on the labels that have a "+"
at the end, but doing it earlier made more sense since we don't care
about the other lines anyway.

* Add some comments

* Don't add review comments for the same problems

Identifying the review comments that the "bot" left previously is a bit
tricky.  First, we fetch all the comments on a PR and ignore the ones
that are outdated (identified by having a null position).

This part is a bit hacky and I haven't thoroughly tested it yet - we
also ignore the comments that weren't made by a login equal to the
BOT_NAME environment variable.

This means that the environment variables for the token and the name
should be for the same account.

Now that we have only the comments that our review bot left, we can
determine if we have any new additions to make.  A new addition is an
addition which is on a new path/position combo and that has a new body.

* Use correct body

The body in the stringified object is the comment text

* Make sure Sentry & ESlint comments are different

If the comments are identical and someone had a call to Sentry and then
an inline comment at the end which disabled something in ESLint, then it
would create a race condition.  This kind of thing is unlikely to
happen, but we should try to guard against it.

* Re-enable eslint check

* Cast position to be int

The double quotes around the int was throwing off the stringified
comparison

* Try to fix bug

* Move function up top for readability

* Exit if this run isn't for a PR

* Remove unused install-dependencies job

* Don't destructure arguments

This doesn't really benefit us here

* Shift boundaries a bit

* Remove old step

This was mostly for a sanity check

* Don't exit early if additions is empty

I want to see if a review comment can be made if the comments array is
empty

* Add length check back in

* Make beginning section a bit more readable

* Refactor to create getPRbotComments function

* Do some functional refactoring

* Add a Sentry log back for testing

* Do some more functional refactoring

* Add some comments

* Fix the "error" handling

* Undo test changes

* Modify circle config

* Update the lockfile

* Try to tag frontend-review-group

* Try to create a review request

* Undo some testing changes

* Refactor labelAdditions a bit

I'm trying to make this more functional to be consistent with the rest
of the code

* Fix typo in comment

Co-Authored-By: Eugene Doan <1067024+U-DON@users.noreply.github.com>

* Rewrite comment for clarity & rename func

* Improve readability of findPattern

* Rename arg for clarity

* Testing with bot token

* Revert "Testing with bot token"

This reverts commit 27edb258ee179430f9bdc3cac4a3db7aea476447.

* Switch to a reducer

* Trigger Circle

* Change wording to emphasize waiting for a review

* Add initial draft for icon-check job

This is for accessibility reasons.  FontAwesome icons typically use the
<i> tag to add icons, and sometimes icons are used to convey meaning
that isn't captured in text.  This is what we want to be on the lookout
for.  In the cases where icons are used with a semantic purpose in mind,
we want to make sure that those semantics are conveyed to a screen
reader.

* Reword comment string to avoid yaml error

* Add another icon as a test

* Be more specific with icon regex

* Add another icon

* Remove icon additions

These were for testing

Co-authored-by: Eugene Doan <1067024+U-DON@users.noreply.github.com>