Remove moment from validateDependentDate() func
This is a bit tricky - the function was using `_.get()` to try to access
a var that didn't exist, so the default formData object was always used
instead.  When moment() is called with any non-date object as its
parameter, it defaults to now.  So, essentially the `dob` variable was
always being assigned to the current datetime.