Use the useLocation hook in Profile2Wrapper
So the component can easily see which route we are currently on and it
can update the breadcrumbs accordingly

Using the useLocation hook broke the existing unit tests that used
shallow mounting. I ran into errors using enzyme's full mount as well.
Specifically the useEffect hook used in the Profile2Wrapper's child
ProfileSideNav was causing enzyme.mount() to blow up. But the component
mounts just fine with React Testing Library.

Full error I saw when trying to mount with enzyme:
```
Error: An error was thrown inside one of your components, but React doesn't know what it was. This is likely due to browser flakiness. React does its best to preserve the "Pause on exceptions" behavior of the DevTools, which requires some DEV-mode only tricks. It's possible that these don't work in your browser. Try triggering the error in production mode, or switching to a modern browser. If you suspect that this is actually an issue with React, please file an issue.
```