Pass the items schema when expanding entity references
Ran into a nasty bug where this created a 2D array because both
expandEntityReferenceArray and transformFields were returning arrays.
This was happening becuase I was passing the schema for the array to
transformFields from expandEntityReferenceArray, but I wanted to pass
the schema for each item when I'm iterating over each item. Because that
makes sense.

There are some times that the output schema doesn't specify an array,
though, so we still want to pass the outputFieldSchema as-is (not
.items). So I had to handle both.