Fix logic for pointer-events on active children
The logic in the CSS for rendering pointer-events:none to disable links on the currently-selected page was very complex and broken. We were cascading "pointer-events:none" down to all children of the active node.

I moved the logic into JS, wrote a unit test to cover it, and removed it from CSS.

https://github.com/department-of-veterans-affairs/va.gov-team/issues/13621