Fix the endless loop and clean up dirtyConsole handling
When we run the content watch, Silversmith is called multiple times.
This is a problem because the process.stdout.write and
process.stderr.write overloads were causing an endless loop.

I cleaned this up by moving the logic to its own module and checking to
see if we've already overloaded the global process functions.