Make new isVAPatient selector
This is an alternate way to see if a user has VA health care. It uses data that already exists on the user model so it does not require making a separate call to ESR to check enrollment status.

This PR also renames the existing selector: `isEnrolledInVAHealthCare` => `isEnrolledInESR`

fix broken import