only trust keepalive response if session-alive is present
In some cases, the response returned to us from the /keepalive endpoint
won't have valid response headers.  One such case is when the browser
fails to submit a "Origin" request header.  When that happens, the
"sesion-alive" response header will be absent, so we shouldn't assume
that the user is logged out.  When missing we should assert that we can't
determine a user's SSOe status and therefore shouldn't try to keep it "in
sync" with their www.va.gov session.