Fix timezone issues with dates in staging and production
Hypothesis: the moment "unix" function converts UNIX timestamp into a UTC date.

Our dateFromUnix filter passes that date to the format function, but without specifying
a timezone, so it uses the local timezone by default, which may be different in
staging and prod.

This PR adds a specific timezone, which should resolve the bug.