WIP. Adding a proper axeConfig object to one test.
I moved away from the previous custom configuration object. It was
going to be too unwieldy to create a custom API and ask our dev teams
to keep it in sync. I went back to the previous example of destructuring
the _13647Exception from the tempOptions {} and used Object.assign to
merge a base axeConfig with any overrides passed into tempOptions.