feature/16544axe-4-bump - Bumping axe to current version 4.1.1 in CI/CD pipeline (#16056)
* Initial work for axe-core configurable test.

* Bump cypress-axe to 0.12 latest

Ensuring the updated axeCheck function behaves correctly
with the newest version of the cypress-axe wrapper. The
following tests were done:
* true negative - Color contrast check added, no errors
* true positive - Color contrast check logged a valid error
* false positive - Select background images were correctly ignored
* false negative - Color contrast error was ignored by_13647Exception

* Bumping Cypress to 6.3.0

Preliminary testing with Cypress 6.3.0 and the modified
axeCheck function. Tested for true positive (a valid color
contrast error was detected) and true negative (color
contrast rule was enabled, no errors) before doing more
exhaustive testing for false positives and false negatives.

* WIP. Adding a proper axeConfig object to one test.

I moved away from the previous custom configuration object. It was
going to be too unwieldy to create a custom API and ask our dev teams
to keep it in sync. I went back to the previous example of destructuring
the _13647Exception from the tempOptions {} and used Object.assign to
merge a base axeConfig with any overrides passed into tempOptions.

* WIP. Bumping to Cypress 6.4

Bumping to the latest version of Cypress and re-running tests to
ensure no regression using Cypress 6.4 with Cypress-axe 0.12 and
axe-core 4.1.1.

* Removed extra rulesets and looping test

This should be the next-to-final step before issing a PR. Latest master
is pulled into the branch and Cypress is on 6.4 Verified tests are
working as expected and that I can pass in different rulesets using
the axe-core configuration object. Remaining items:

* Test static content
* Write documentation

* Final tests run for static content

I made sure the Nightwatch tests for static content were run against
the latest Drupal content, using `yarn build:content --pull-drupal`
on my feature branch. I ran the tests twice for all static content
with no errors, then ran the Nightwatch e2e suite for apps as a
secondary check.

* Fixing a broken unit test in ProgressBar

* Fixed the broken unit test
* Updated the FileField.jsx to use the DS version of ProgressBar
* Reverted the Facility Locator gaEvents Cypress e2e test (removing it
  from this PR)
* Retested unit tests locally to ensure no further regression

* Removing legacy ProgressBar component

After consulting with the Design System team, we felt it was best
to remove the legacy version of the ProgressBar component and the
related unit test. I linked to the new design system version of
the ProgressBar in both use cases. All unit tests passing.