Hide/show search this area button correctly
https://github.com/department-of-veterans-affairs/va.gov-team/issues/20099

This was broken because we were calling activateSearchAreaControl before the map zoom
was finished, so we were always one zoom event behind.
Calling activateSearchAreaControl on zoomend/dragend fixes this bug, and also
triggers the activateSearchAreaControl logic when zooming using the mouse wheel
or any other method, not just the zoom in/out buttons.