Update Applicant Information Form; Add Review Box (#18099)
* Update Applicant Information Form; Add Review Box

Add the Date of Birth field and remove the SSN field. Additionally,
update the form title and subtitle.

This also adds a new form paradigm: the Review Box (for lack of a
better name). This is meant to display pre-populated data for the
veteran to review with the option to edit the information. The
primary purpose of this new read-first, edit-later paradigm is to
provide a better review experience and discourage easy editing for
fields that most likely shouldn't change, like name and birthday.

* Add Additional Form Validation

Co-authored-by: nbuckafs <nathaniel.buck@accenturefederal.com>