Initial routing setup (#40)
* Add basic Page component

initialValues is not great. I'll work on a Router component next and
have initialValues come from a React context.

* Setup a simple Router component

* Remove Routable type

We can add this back when we add Chapter

* Conditionally put a Route in a Page

In our simple form example, the Page isn't inside a Router, and that
throws an error. We will need some kind of guard to alow Pages to be
used without formulate's routing submodule

* Revert "Conditionally put a Route in a Page"

This reverts commit f06372f0226a3823fc546f0d0f1b0e2cca1ccfb4.

* Remove Page from simple-form example

Page is for routing only, which is for more complex examples

* Remove Page import

* Inline the example

* Move RouterProps to types file

* Remove placeholder prev/next links

* Remove hard-coded initialValues

This will come back once we add a React Context

* Add test for Page component

* Add test for Router

Router and Page are tightly coupled due to typing, so the Router test
uses the Page as well

* Add some documentation

* Update test/routing/Page.test.tsx

Co-authored-by: Chris Valarida <chris.valarida@adhocteam.us>

* Add comment explaining why the empty Route is last

* Add a Routable type

* Change PaheType's children to be single or multi

* Update generated docs

* Include routing types in export & docs

* Move PageProps into types file

* Update src/routing/types.ts

Co-authored-by: Chris Valarida <chris.valarida@adhocteam.us>

* Add descriptions to the types

* Remove PageType

Co-authored-by: Chris Valarida <chris.valarida@adhocteam.us>