Checkbox is no longer broken (#19738)
* Checkbox is no longer broken

It can be checked and unchecked now.

* Checkbox is no longer broken

Turns out checkbox lowercases name for the target name. Camelcased names will always be false in checkboxes.

* Selector/calculator.js uses the same input prop

Changed "outsideUS" property used in the checkbox here as well to ensure that the calculator is using the same input property.

Everything appears to be working as expected.